requirements.txt
apache-airflow-providers-apache-spark==5.3.2
apache-airflow-providers-openlineage==2.7.0
pyspark==3.5.6
numpy<2
pandas==2.3.2
urllib3==2.5.0
boto3==1.40.33
sodapy==2.2.0